Finished up environments
Finished up environments (for now) by moving it off into its own module
