+++ include/environments.h
@@ -1,0 +1,14 @@
+/* environments.h */
+
+#ifndef FILE_ENVIRONMENTS_SEEN
+#define FILE_ENVIRONMENTS_SEEN
+
+extern object *lookup_variable_value(object *var, object *env);
+extern void set_variable_value(object *var, object *val, object *env);
+extern void define_variable(object *var, object *val, object *env);
+extern void init_environments();
+
+extern object *the_empty_environment;
+extern object *the_global_environment;
+
+#endif // !FILE_ENVIRONMENTS_SEEN
